add confirmation to delete product

+++ client/src/components/homePageComponents/stock/StockSearch.jsx
@@ -116,6 +116,10 @@ const StockSearch = () => {
         setIsButtonLoading(true);
         setDeleteId(id);
 
+        const confirm = window.confirm('Are You sure to want to delete this product? This Action cannot be undone!')
+
+        if(!confirm) return
+
         try {
             const response = await config.deleteProduct(id);
             if (response) {
